microservices/registration_service/requirements.txt
# Django and core dependencies
Django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.1
django-filter==23.3

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.7
django-redis==5.4.0

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# HTTP requests and API calls
requests==2.31.0

# Email and notifications
django-anymail==10.2

# Validation and utilities
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
email-validator==2.1.0
validators==0.22.0

# Security
cryptography==41.0.7
django-ratelimit==4.1.0
django-axes==6.1.1

# Monitoring and logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Environment and configuration
python-decouple==3.8

# Timezone handling
pytz==2023.3

# File handling
Pillow==10.1.0

# Deployment
gunicorn==21.2.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
django-stubs==4.2.7

# Kenya-specific utilities
pycountry==23.12.11
babel==2.13.1

# Additional file handling
django-storages==1.14.2

# JSON handling
orjson==3.9.10

# Date utilities
python-dateutil==2.8.2

# Workflow and state management
django-fsm==2.8.1

# Caching
django-cachalot==2.6.1

# Admin enhancements
django-admin-interface==0.28.6
django-colorfield==0.10.1

# Export functionality
django-import-export==3.3.1
openpyxl==3.1.2

# Web scraping (for institutional verification)
beautifulsoup4==4.12.2
lxml==4.9.3

# Machine learning (for risk assessment)
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.3

# Text processing
nltk==3.8.1
textblob==0.17.1

# Geolocation
geopy==2.4.0

# Configuration management
pyyaml==6.0.1

# Networking
requests-oauthlib==1.3.1

# Debugging
django-debug-toolbar==4.2.0

# Testing utilities
responses==0.24.1
requests-mock==1.11.0

# Security scanning
bandit==1.7.5
safety==2.3.5

# WebSocket support (for real-time notifications)
channels==4.0.0
channels-redis==4.1.0

# Task scheduling
django-crontab==0.7.1
celery-beat==2.5.0

# Model utilities
django-model-utils==4.3.1

# Form utilities
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Audit logging
django-auditlog==2.3.0

# Version control
django-reversion==5.0.8

# Utilities
django-braces==1.15.0
django-guardian==2.4.0

# Country utilities
django-countries==7.5.1

# Quality assurance utilities
coverage==7.3.2

# Testing utilities
pytest-xdist==3.5.0

# Communication utilities
twilio==8.10.3